♻️ Refactored embeded -> embededCreator
