+++ src/interfaces/Embeded.ts
@@ -1,0 +1,4 @@
+export interface EmbededField {
+  name: string;
+  value: string;
+}
